+++ examples/space.c
@@ -10,7 +10,7 @@ double mtime(){
 
 int main(int argc, char **argv){
   int exit = 1;
-  size_t samples = 100;
+  uint32_t samples = 100;
   WINDOW *window = 0;
   struct mixed_segment_sequence sequence = {0};
   struct mixed_segment space = {0}, downmix = {0};
