+++ src/buffer.c
@@ -1,6 +1,6 @@
 #include "internal.h"
 
-MIXED_EXPORT int mixed_make_buffer(size_t size, struct mixed_buffer *buffer){
+MIXED_EXPORT int mixed_make_buffer(uint32_t size, struct mixed_buffer *buffer){
   mixed_err(MIXED_NO_ERROR);
   if(buffer->_data && !buffer->virtual){
     mixed_err(MIXED_BUFFER_ALLOCATED);
@@ -35,18 +35,18 @@ MIXED_EXPORT int mixed_buffer_clear(struct mixed_buffer *buffer){
   return 1;
 }
 
-static inline size_t free_for_r2(struct mixed_buffer *buffer){
+static inline uint32_t free_for_r2(struct mixed_buffer *buffer){
   return buffer->r1_start - buffer->r2_start - buffer->r2_size;
 }
 
-static inline size_t free_after_r1(struct mixed_buffer *buffer){
+static inline uint32_t free_after_r1(struct mixed_buffer *buffer){
   return buffer->size - buffer->r1_start - buffer->r1_size;
 }
 
-MIXED_EXPORT int mixed_buffer_request_write(float **area, size_t *size, struct mixed_buffer *buffer){
+MIXED_EXPORT int mixed_buffer_request_write(float **area, uint32_t *size, struct mixed_buffer *buffer){
   mixed_err(MIXED_NO_ERROR);
   if(buffer->r2_size){
-    size_t free = free_for_r2(buffer);
+    uint32_t free = free_for_r2(buffer);
     if(*size < free) free = *size;
     if(free == 0){
       *size = 0;
@@ -58,7 +58,7 @@ MIXED_EXPORT int mixed_buffer_request_write(float **area, size_t *size, struct m
     buffer->reserved_start = buffer->r2_start + buffer->r2_size;
     *area = buffer->_data + buffer->reserved_start;
   }else{
-    size_t free = free_after_r1(buffer);
+    uint32_t free = free_after_r1(buffer);
     if(buffer->r1_start <= free){
       if(free == 0){
         *size = 0;
@@ -85,7 +85,7 @@ MIXED_EXPORT int mixed_buffer_request_write(float **area, size_t *size, struct m
   return 1;
 }
 
-MIXED_EXPORT int mixed_buffer_finish_write(size_t size, struct mixed_buffer *buffer){
+MIXED_EXPORT int mixed_buffer_finish_write(uint32_t size, struct mixed_buffer *buffer){
   mixed_err(MIXED_NO_ERROR);
   if(buffer->reserved_size < size){
     mixed_err(MIXED_BUFFER_OVERCOMMIT);
@@ -105,7 +105,7 @@ MIXED_EXPORT int mixed_buffer_finish_write(size_t size, struct mixed_buffer *buf
   return 1;
 }
 
-MIXED_EXPORT int mixed_buffer_request_read(float **area, size_t *size, struct mixed_buffer *buffer){
+MIXED_EXPORT int mixed_buffer_request_read(float **area, uint32_t *size, struct mixed_buffer *buffer){
   mixed_err(MIXED_NO_ERROR);
   if(buffer->r1_size == 0){
     *size = 0;
@@ -117,15 +117,15 @@ MIXED_EXPORT int mixed_buffer_request_read(float **area, size_t *size, struct mi
   return 1;
 }
 
-MIXED_EXPORT size_t mixed_buffer_available_read(struct mixed_buffer *buffer){
+MIXED_EXPORT uint32_t mixed_buffer_available_read(struct mixed_buffer *buffer){
   return buffer->r1_size;
 }
 
-MIXED_EXPORT size_t mixed_buffer_available_write(struct mixed_buffer *buffer){
+MIXED_EXPORT uint32_t mixed_buffer_available_write(struct mixed_buffer *buffer){
   return (buffer->r2_size)? free_for_r2(buffer) : free_after_r1(buffer);
 }
 
-MIXED_EXPORT int mixed_buffer_finish_read(size_t size, struct mixed_buffer *buffer){
+MIXED_EXPORT int mixed_buffer_finish_read(uint32_t size, struct mixed_buffer *buffer){
   mixed_err(MIXED_NO_ERROR);
   if(buffer->r1_size < size){
     mixed_err(MIXED_BUFFER_OVERCOMMIT);
@@ -147,7 +147,7 @@ MIXED_EXPORT int mixed_buffer_transfer(struct mixed_buffer *from, struct mixed_b
   mixed_err(MIXED_NO_ERROR);
   if(from != to){
     float *read, *write;
-    size_t samples = SIZE_MAX;
+    uint32_t samples = UINT32_MAX;
     mixed_buffer_request_read(&read, &samples, from);
     mixed_buffer_request_write(&write, &samples, to);
     memcpy(write, read, sizeof(float)*samples);
@@ -161,7 +161,7 @@ MIXED_EXPORT int mixed_buffer_copy(struct mixed_buffer *from, struct mixed_buffe
   mixed_err(MIXED_NO_ERROR);
   if(from != to){
     float *read, *write;
-    size_t samples = SIZE_MAX;
+    uint32_t samples = UINT32_MAX;
     mixed_buffer_request_read(&read, &samples, from);
     mixed_buffer_request_write(&write, &samples, to);
     memcpy(write, read, sizeof(float)*samples);
@@ -170,7 +170,7 @@ MIXED_EXPORT int mixed_buffer_copy(struct mixed_buffer *from, struct mixed_buffe
   return 1;
 }
 
-MIXED_EXPORT int mixed_buffer_resize(size_t size, struct mixed_buffer *buffer){
+MIXED_EXPORT int mixed_buffer_resize(uint32_t size, struct mixed_buffer *buffer){
   mixed_err(MIXED_NO_ERROR);
   // FIXME: fixup
   void *new = realloc(buffer->_data, size*sizeof(float));
