+++ src/segments/null.c
@@ -13,7 +13,7 @@ int null_segment_start(struct mixed_segment *segment){
   return 1;
 }
 
-int null_segment_set(size_t field, size_t location, void *buffer, struct mixed_segment *segment){
+int null_segment_set(uint32_t field, uint32_t location, void *buffer, struct mixed_segment *segment){
   switch(field){
   case MIXED_BUFFER:
     switch(location){
@@ -29,7 +29,7 @@ int null_segment_set(size_t field, size_t location, void *buffer, struct mixed_s
 int void_segment_mix(struct mixed_segment *segment){
   struct mixed_buffer *data = (struct mixed_buffer *)segment->data;
   float *buffer;
-  size_t frames = SIZE_MAX;
+  uint32_t frames = UINT32_MAX;
   mixed_buffer_request_read(&buffer, &frames, data);
   mixed_buffer_finish_read(frames, data);
   return 1;
@@ -38,7 +38,7 @@ int void_segment_mix(struct mixed_segment *segment){
 int zero_segment_mix(struct mixed_segment *segment){
   struct mixed_buffer *data = (struct mixed_buffer *)segment->data;
   float *buffer;
-  size_t frames = SIZE_MAX;
+  uint32_t frames = UINT32_MAX;
   mixed_buffer_request_write(&buffer, &frames, data);
   memset(buffer, 0, frames*sizeof(float));
   mixed_buffer_finish_write(frames, data);
